test: split service handler tests by domain
